+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 -   This project adheres to [Semantic Versioning][semver].
 -   This project uses [Gitmoji][gitmoji]
 
+## 3.1.0
+
+-   :building_construction: `format` now takes a template literals.
+-   :building_construction: `mix` now takes a template literals.
+
 ## 3.0.5
 
 -   :arrow_up: dependency upgrade.